+++ src/components/SignupForm/hooks.js
@@ -1,4 +1,4 @@
-import { useInput } from "../../components/@shared/Input/hooks";
+import { useInput } from "../@shared/Input/hooks";
 import { isValidAge, isValidEmail, isValidPassword } from "./validator";
 
 export const useSignupEmail = () => useInput(isValidEmail);
