+++ src/components/StationsForm/hooks.js
@@ -1,4 +1,4 @@
-import { useInput } from "../../components/@shared/Input/hooks";
+import { useInput } from "../@shared/Input/hooks";
 import isValidStationName from "./validator";
 
 export const useStationName = () =>
